+++ saga/lib/tokenizers/saga_tokenizers.ml
@@ -3,6 +3,7 @@
   SPDX-License-Identifier: ISC
   ---------------------------------------------------------------------------*)
 
+module Grapheme = Grapheme
 module Unicode = Unicode
 module Normalizers = Normalizers
 module Pre_tokenizers = Pre_tokenizers